requirements.txt
aiohttp
Flask
langdetect
numpy
openai
pdf2image
Pillow
pytesseract
python-dotenv
reportlab
svglib
Werkzeug
rlPyCairo